Add length(::Channel) and make isready/isempty atomic
This removes the data race from isready() and isempty(), which are now
implemented in terms of length(). A new atomic `length` field is added to
track the channel length separately from the buffer and wait queue -
these consist of Vectors which cannot easily have their length fields
read and written atomically.

Viewing Channel as a container, n_avail(::Channel) is now renamed
length().

For buffered channels, the length now includes a count of any waiting
tasks in addition to the number of buffered items. This makes it
consistent with the length computation for unbuffered channels.